pyproject.toml
[tool.poetry]
name = "seiright"
version = "0.0.1"
description = "LLM to check financial institution compliance"
authors = ["Mohit Sharma"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.1"
numpy = "^1.26.4"
scipy = "^1.12.0"
matplotlib = "^3.8.3"
fastapi = "^0.115.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.12"
bcrypt = "4.0.1"
openai = "^1.52.0"
anthropic = "^0.36.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-mock = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
